+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased](https://github.com/wilfredinni/noodle/tree/development)
 
+### Fixed
+
+- Fixed all options printing the help when a Command has no user-defined arguments.
+
+### Added
+
+- Added `TooManyArguments` warning.
+
 ## [0.0.2](https://github.com/wilfredinni/noodle/releases/tag/0.0.2) - 2019-05-15
 
 Initial release.